backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-jazzmin==3.0.0
djangorestframework==3.15.2
geojson==3.1.0
google-api-core==2.20.0
google-auth==2.35.0
googleapis-common-protos==1.65.0
idna==3.10
imagesize==1.4.1
labelbox==5.1.0
mypy==1.10.1
mypy-extensions==1.0.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyGeoTile==1.0.6
pyproj==3.6.1
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
shapely==2.0.6
six==1.16.0
sqlparse==0.5.1
StrEnum==0.4.15
tqdm==4.66.5
typeguard==4.3.0
typing_extensions==4.12.2
urllib3==2.2.3
